align with context
The phrase mentions a subset of the capabilities of RHAMT and exemplifies Websphere and Weblogic. In this context "containers" is confusing. 

The first idea is "Application servers"

RHAMT however also supports other sources than application servers. 

I'd like to use "footprints" as a general term but then you often mean the size of a deployment. 

How about "technologies" ?